use object caching when available for json ld caching (and others) #626

+++ src/includes/class-wordlift-abstract-cache-service.php
@@ -48,6 +48,15 @@ abstract class Wordlift_Abstract_Cache_Service {
 	private $type;
 
 	/**
+	 * A string holding the cache key prefix valuet.
+	 *
+	 * @since  3.16.0
+	 * @access private
+	 * @var string
+	 */
+	private $cache_prefix;
+
+	/**
 	 * Wordlift_Cache_Service constructor.
 	 *
 	 * @param string $type   A string identifying the group of item being cached
@@ -64,6 +73,40 @@ abstract class Wordlift_Abstract_Cache_Service {
 		if ( WP_DEBUG && ( $this->type !== $type ) ) {
 			throw new Exception( "type is not simple lowercase string type=$type" );
 		}
+
+		// If there is an object caching plugin, we are utilize the caching APIs
+		// with them the only way to flush a cache is to not access it any longer.
+		// For that we are going to use a prefix value used for our keys which will be
+		// set to a different value on each flush and stored by itself in the cache.
+		if ( wp_using_ext_object_cache() ) {
+			$this->cache_prefix = wp_cache_get( $type, 'wordlift_prefix_key' );
+			if ( ! $this->cache_prefix ) {
+				$this->set_new_cache_prefix();
+			}
+		}
+	}
+
+	/**
+	 * Generate a new cache prefix key
+	 *
+	 * @since 3.16.0
+	 */
+	private function set_new_cache_prefix() {
+		$new_key = rand( 1, 10000000000 );
+		wp_cache_set( $this->type, $new_key, 'wordlift_prefix_key', 0 );
+		$this->cache_prefix = $new_key;
+	}
+
+	/**
+	 * Generate a cache key which can be unique across cache flushes. Utilizes
+	 * the fact we prevent the use of "-" in ids.
+	 *
+	 * @since 3.16.0
+	 *
+	 * @param string $id The identifier of the specific data.
+	 */
+	private function get_cache_key( $id ) {
+		return $this->cache_prefix . '-' . $id;
 	}
 
 	/**
@@ -110,27 +153,31 @@ abstract class Wordlift_Abstract_Cache_Service {
 	 */
 	public function get( $id, $default = false ) {
 
-		$filename = $this->get_filename( $id );
+		// If there is an object caching plugin, we are utilize the caching APIs.
+		if ( wp_using_ext_object_cache() ) {
+			return wp_cache_get( $this->get_cache_key( $id ), 'wordlift' );
+		} else {
+			$filename = $this->get_filename( $id );
 
-		if ( ! file_exists( $filename ) ) {
-			return $default;
+			if ( ! file_exists( $filename ) ) {
+				return $default;
+			}
+
+			$content = file_get_contents( $filename );
+
+			// Content consist of a a json formatted
+			// array containing the expired time and the value.
+			$content = json_decode( $content, true );
+
+			if ( is_null( $content ) ) { // garbage found instead of proper json.
+				return $default;
+			}
+
+			if ( $content['expire'] > time() ) { // If expired.
+				return $default;
+			}
+			return $content['value'];
 		}
-
-		$content = file_get_contents( $filename );
-
-		// Content consist of a a json formatted
-		// array containing the expired time and the value.
-		$content = json_decode( $content, true );
-
-		if ( is_null( $content ) ) { // garbage found instead of proper json.
-			return $default;
-		}
-
-		if ( $content['expire'] > time() ) { // If expired.
-			return $default;
-		}
-
-		return $content['value'];
 	}
 
 	/**
@@ -145,16 +192,21 @@ abstract class Wordlift_Abstract_Cache_Service {
 	 */
 	public function set( $id, $value, $expiry ) {
 
-		$filename = $this->get_filename( $id );
+		// If there is an object caching plugin, we are utilize the caching APIs.
+		if ( wp_using_ext_object_cache() ) {
+			wp_cache_set( $this->get_cache_key( $id ), $value, 'wordlift' , $expiry );
+		} else {
+			$filename = $this->get_filename( $id );
 
-		// Create the content saved in the file. It consist of a json of
-		// array containing the expired time and the value.
-		$content = array(
-			'expire' => (0 === $expiry) ? 0 : time() + $expiry,
-			'value' => $value,
-		);
+			// Create the content saved in the file. It consist of a json of
+			// array containing the expired time and the value.
+			$content = array(
+				'expire' => (0 === $expiry) ? 0 : time() + $expiry,
+				'value' => $value,
+			);
 
-		file_put_contents( $filename, wp_json_encode( $content ) );
+			file_put_contents( $filename, wp_json_encode( $content ) );
+		}
 	}
 
 	/**
@@ -166,13 +218,18 @@ abstract class Wordlift_Abstract_Cache_Service {
 	 */
 	public function delete( $id ) {
 
-		$filename = $this->get_filename( $id );
+		// If there is an object caching plugin, we are utilize the caching APIs.
+		if ( wp_using_ext_object_cache() ) {
+			wp_cache_delete( $this->get_cache_key( $id ), 'wordlift' );
+		} else {
+			$filename = $this->get_filename( $id );
 
-		if ( ! file_exists( $filename ) ) {
-			return;
+			if ( ! file_exists( $filename ) ) {
+				return;
+			}
+
+			unlink( $filename );
 		}
-
-		unlink( $filename );
 	}
 
 	/**
@@ -182,17 +239,24 @@ abstract class Wordlift_Abstract_Cache_Service {
 	 */
 	public function flush() {
 
-		$dir = $this->get_directory_name();
+		// If there is an object caching plugin, we can not actually flush the values
+		// in any reliable way so we just change the prefix and let the caching utility
+		// garbage collect the old instances while we use new ones.
+		if ( wp_using_ext_object_cache() ) {
+			$this->set_new_cache_prefix();
+		} else {
+			$dir = $this->get_directory_name();
 
-		if ( ! file_exists( $dir ) ) {
-			return;
-		}
+			if ( ! file_exists( $dir ) ) {
+				return;
+			}
 
-		$files = glob( "$dir/*" . self::FILE_SUFFIX );
-		if ( ! empty( $files ) ) {
-			foreach ( $files as $file ) { // iterate files.
-				if ( is_file( $file ) ) {
-					unlink( $file ); // delete file.
+			$files = glob( "$dir/*" . self::FILE_SUFFIX );
+			if ( ! empty( $files ) ) {
+				foreach ( $files as $file ) { // iterate files.
+					if ( is_file( $file ) ) {
+						unlink( $file ); // delete file.
+					}
 				}
 			}
 		}
